MOB-25686 Add functionality to take receipt capture from ExpenseIt list MOB-25685 [Android] Add new UI for ExpenseIt capture from ExpenseIt list
Description:
Added Ui and functionality and take a receipt image from expenseList.
Currently the UI works with 2 button option. Manual for manual entry and ExpenseIt for image Capture. On Manual Entry that should go to the same old edit details while the new Receipt capture should take to the  image capture that will get uploaded to expenseIt.
Once uploaded, the list will be refreshed to show the changes.
Reviewed By: